requirements.txt
pytest
pytest-html
pytest-xdist
pytest-ordering
allure-pytest
pytest-refunfailures